Prototype Finished

+++ Scenes/Global/LevelStats.cs
@@ -4,5 +4,5 @@ using System;
 public partial class LevelStats : Node
 {
 	[Export] public PackedScene Level;
-	[Export] public float Time;
+	[Export] public double Time;
 }
